refactor: move fetchData to functions
