+++ src/config.ts
@@ -1,0 +1,2 @@
+export const GARMIN_APP_VERSION = '4.55.3.1'
+export const LOGIN_DELAY = 1100
