+++ src/index.ts
@@ -2,19 +2,17 @@ import { Command } from 'commander'
 import dayjs from 'dayjs'
 import dotenv from 'dotenv'
 import fs from 'fs/promises'
-import { chromium } from 'playwright-chromium'
 import { reverse, sortBy, uniqWith } from 'rambda'
 import { version } from '../package.json'
+import { fetchData } from './functions'
 
 import type { GarminCommandOptions, GarminDataItem } from './types'
 
 dotenv.config()
 
-const GARMIN_APP_VERSION = '4.55.3.1'
-const LOGIN_DELAY = 1100
 let data: GarminDataItem[] = []
 let items: GarminDataItem[] = []
-let itemsOriginally = 0
+let existingActivitiesCount = 0
 let DEBUG = false
 let [searchYear, searchMonth] = dayjs().format('YYYY-M').split('-')
 let browserStoragePath = 'sessionStorage.json'
@@ -35,156 +33,7 @@ program
 
 const progOptions = program.opts() as GarminCommandOptions
 
-let forceAuth = !!progOptions.authenticate
-
-function sleep(ms: number) {
-  return new Promise<void>((resolve) => {
-    setTimeout(resolve, ms)
-  })
-}
-
-async function fetchData(year: string, month: string) {
-  return new Promise<GarminDataItem[]>(async (resolve, reject) => {
-    month = `${parseInt(month) - 1}`
-    let context = null
-
-    const browser = await chromium.launch({
-      args: ['--disable-dev-shm-usage'],
-      headless: true,
-    })
-
-    if (!forceAuth) {
-      try {
-        const storageData = await fs.readFile(browserStoragePath, {
-          encoding: 'utf8',
-        })
-        const storageState = JSON.parse(storageData)
-        if (DEBUG) {
-          console.log('session storage found: ', storageState)
-        }
-        context = await browser.newContext({
-          storageState,
-        })
-        console.log(`✓ Using existing browser session.`)
-      } catch (err) {
-        console.log(`✓ Existing browser session not found.`)
-        forceAuth = true
-      }
-    }
-
-    if (forceAuth || !context) {
-      context = await browser.newContext({
-        userAgent:
-          'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/102.0.0.0 Safari/537.36',
-        viewport: { width: 1280, height: 1024 },
-      })
-    }
-
-    const page = await context.newPage()
-    page.setExtraHTTPHeaders({
-      'X-app-ver': GARMIN_APP_VERSION,
-      'NK': 'NT',
-      'accept-language': 'en-GB,en-US;q=0.9,en;q=0.8',
-    })
-
-    const url = `https://connect.garmin.com/modern/proxy/calendar-service/year/${year}/month/${month}`
-    if (DEBUG) {
-      console.log('fetchData URL: ', url)
-    }
-
-    if (forceAuth) {
-      try {
-        await page.goto('https://connect.garmin.com/signin')
-        await page.waitForSelector('iframe')
-        await sleep(LOGIN_DELAY)
-
-        try {
-          await page.waitForSelector('#truste-consent-button')
-          await page.click('#truste-consent-button')
-          console.log('#truste-consent-button clicked')
-        } catch {
-          console.log('#truste-consent-button never came')
-        }
-
-        if (DEBUG) {
-          await sleep(LOGIN_DELAY)
-          await page.screenshot({
-            path: `debug-00-unfilled-login.png`,
-            fullPage: true,
-          })
-        }
-
-        await page.frames()[1].check('#login-remember-checkbox')
-        await page.frames()[1].fill('input[name="username"]', process.env.GARMIN_CONNECT_USERNAME!)
-        await page.frames()[1].fill('input[name="password"]', process.env.GARMIN_CONNECT_PASSWORD!)
-
-        if (DEBUG) {
-          await sleep(LOGIN_DELAY)
-          await page.screenshot({
-            path: `debug-01-filled-login.png`,
-            fullPage: true,
-          })
-        }
-
-        await page.frames()[1].click('#login-btn-signin')
-        await page.waitForSelector('.user-profile')
-        await sleep(LOGIN_DELAY * 2)
-        if (DEBUG) {
-          await page.screenshot({
-            path: `debug-02-after-login.png`,
-            fullPage: true,
-          })
-        }
-
-        await page.goto('https://connect.garmin.com/modern/calendar')
-
-        const storage = await context.storageState()
-        const storageJson = JSON.stringify(storage, null, 2)
-        if (DEBUG) {
-          console.log('session storage: ', storageJson)
-        }
-        await fs.writeFile(browserStoragePath, storageJson)
-        console.log(`✓ Browser session created and saved to ${browserStoragePath}`)
-      } catch (err) {
-        if (DEBUG) {
-          console.log(err)
-        }
-        return reject(err)
-      }
-    }
-
-    await page.goto('https://connect.garmin.com/modern/calendar')
-    await sleep(LOGIN_DELAY * 2)
-    if (DEBUG) {
-      await page.screenshot({
-        path: `debug-03-calendar-home.png`,
-        fullPage: true,
-      })
-    }
-
-    page
-      .goto(url)
-      .then(async (response) => {
-        if (!response) {
-          return reject(new Error('Error: no response from Garmin Connect'))
-        }
-        const body = await response.body()
-        const bodyString = body.toString()
-        if (DEBUG) {
-          console.log('Raw data: ', bodyString)
-        }
-
-        const content = await JSON.parse(bodyString)
-        await browser.close()
-        process.stdout.write(` Done.\n`)
-        return resolve(content.calendarItems)
-      })
-      .catch((error) => {
-        console.log('Error fetching data: ', error)
-        return reject(error)
-      })
-  })
-}
+const forceAuth = !!progOptions.authenticate
 
 ;(async () => {
   if (!process.env.GARMIN_CONNECT_USERNAME) {
@@ -209,8 +58,8 @@ async function fetchData(year: string, month: string) {
   try {
     const contents = await fs.readFile(progOptions.outputFile, { encoding: 'utf8' })
     data = JSON.parse(contents) as GarminDataItem[]
-    itemsOriginally = data.length
-    console.log(`✓ Found existing file with ${itemsOriginally} items.`)
+    existingActivitiesCount = data.length
+    console.log(`✓ Found existing file with ${existingActivitiesCount} items.`)
   } catch (err) {
     console.log('No existing file found.')
   }
@@ -218,7 +67,11 @@ async function fetchData(year: string, month: string) {
   process.stdout.write(`Querying ${searchYear}-${searchMonth}.. `)
 
   try {
-    items = await fetchData(searchYear, searchMonth)
+    items = await fetchData(searchYear, searchMonth, {
+      debug: DEBUG,
+      forceAuth,
+      browserStoragePath,
+    })
   } catch (err) {
     console.log('Data fetching failed.')
     process.exit(1)
@@ -252,7 +105,7 @@ async function fetchData(year: string, month: string) {
     }
   }
 
-  if (data.length > itemsOriginally) {
+  if (data.length > existingActivitiesCount) {
     await fs.writeFile(progOptions.outputFile, JSON.stringify(data, null, 2))
     console.log(`Saved ${data.length} items.`)
     process.exit(0)
